Implementando upload de imagens

+++ laravel+vue.js/code/app_locadora_carros/app/Http/Controllers/MarcaController.php
@@ -21,7 +21,7 @@ class MarcaController extends Controller
     {
         //$marcas = Marca::all();
         $marcas = $this->marca->all();
-        return $marcas;
+        return response()->json($marca, 200);
     }
 
     /**
@@ -43,8 +43,12 @@ class MarcaController extends Controller
     public function store(Request $request)
     {
         //$marca = Marca::create($request->all());
+
+        //Encaminhar o 'accept' no header da requisição!
+        $request->validate($this->marca->rules(), $this->marca->feedback());
+
         $marca = $this->marca->create($request->all());
-        return $marca;
+        return response()->json($marca, 201);
     }
 
     /**
@@ -56,7 +60,10 @@ class MarcaController extends Controller
     public function show($id)
     {
         $marca = $this->marca->find($id);
-        return $marca;
+        if ($marca === null) {
+            return response()->json(['erro' => 'Recurso pesquisado não existe.'], 404);
+        }
+        return response()->json($marca, 201);
     }
 
     /**
@@ -81,8 +88,14 @@ class MarcaController extends Controller
     {
         //$marca->update($request->all());
         $marca = $this->marca->find($id);
+
+        if ($marca === null) {
+            return response()->json(['erro' => 'Impossível realizar atualização. O recurso solicitado não existe.'], 404);
+        }
+
+        $request->validate($marca->rules(), $marca->feedback());
         $marca->update($request->all());
-        return $marca;
+        return response()->json($marca, 201);
     }
 
     /**
@@ -94,7 +107,12 @@ class MarcaController extends Controller
     public function destroy($id)
     {
         $marca = $this->marca->find($id);
+
+        if ($marca === null) {
+            return response()->json(['Impossível realizar a exclusão. O recursos solicitado não existe.'], 404);
+        }
+
         $marca->delete();
-        return ['msg' => 'A marca foi removida com sucesso.'];
+        return response()->json(['msg' => 'A marca foi removida com sucesso.'], 200);
     }
 }
